Add regression test for hidden property selector
